move app.py to project root
